requirements.txt
PyJWT
cryptography 
requests
urllib3
flask